refactor(login): created loginForm component
