+++ src/components/index.ts
@@ -2,5 +2,6 @@ import CampaignCard from "@/components/CampaignCard/CampaignCard";
 import Pagination from "@/components/Pagination/Pagination";
 import SearchBar from "@/components/SearchBar/SearchBar";
 import SignupForm from "@/components/SignupForm/SignupForm";
+import LoginForm from "@/components/LoginForm/LoginForm";
 
-export {CampaignCard, Pagination, SearchBar, SignupForm};
+export {CampaignCard, Pagination, SearchBar, SignupForm, LoginForm};
